extract Texture class
