+++ src/BallpointPen.js
@@ -1,6 +1,7 @@
 import { last } from 'lodash';
 import { map } from './utils';
-import { paperColorDark, penR, penG, penB, filterWeight, filterWeightInverse, stepInterval, defaultSample, textureSamplesLength, inkTextureSamples, inkTextureBase64 } from './constants';
+import { paperColorDark, penR, penG, penB, filterWeight, filterWeightInverse, stepInterval, defaultSample } from './constants';
+import {Texture} from './Texture';
 import Point from './Point';
 
 /*
@@ -33,14 +34,12 @@ var canvas;
 var rawStrokes;
 var curRawSampledStroke;
 var curFilteredStroke;
-var textureSampleStep;
 var lastControlPoint;
 var stepOffset;
 var pointCounter;
 var sample;
 
-// Generate Texture Samples
-var textureSampleLocations = getSampleLocations();
+var texture;
 
 // ------------------------------------------
 // Ploma
@@ -58,6 +57,7 @@ export class BallpointPen {
 		ctx = canvas.getContext('2d');
 		ctx.imageSmoothingEnabled = false;
 		sample = defaultSample;
+		texture = new Texture();
 		this.clear();
 	}
 
@@ -85,7 +85,7 @@ export class BallpointPen {
 		lastControlPoint = null;
 		stepOffset = 0.0;
 		pointCounter = 0;
-		textureSampleStep = 0;
+		texture.clear();
 	}
 
 	// ------------------------------------------
@@ -560,61 +560,6 @@ function calculateStepPoints(p0, p1, p2, p3) {
 	return stepPoints;
 }
 
-// ------------------------------------------
-// TEXTURE
-//
-
-function getImageFromBase64(base64, type) {
-	var img = new Image();
-	img.src = 'data:image/' + type + ';base64, ' + base64;
-
-	return img;
-}
-
-function getImageDataFromImage(img) {
-	var canvas = document.createElement('canvas');
-	var ctx = canvas.getContext('2d');
-	canvas.width = img.width;
-	canvas.height = img.height;
-	ctx.drawImage(img, 0, 0);
-	return ctx.getImageData(0, 0, img.width, img.height).data;
-}
-
-function getSampleLocations() {
-	var img = getImageFromBase64(inkTextureBase64, 'jpeg');
-	var imageData = getImageDataFromImage(img);
-	var imageDataGrays = [];
-	var textureOffsetX = 0;
-	var textureOffsetY = 0;
-	var textureSampleLocations = [];
-
-	// Read grays from image
-	for(var i = 0; i < imageData.length; i+=4) {
-		imageDataGrays.push(1 - imageData[i]/255);
-	}
-
-	// Read samples from mirrored-and-tiled grays
-	for (let i = 0; i < textureSamplesLength; i++) {
-		// Get normalized pixel within texture
-		var T_s = textureOffsetX / (img.width - 1);
-		var T_t = textureOffsetY / (img.height - 1);
-		var s = Math.abs(Math.abs(T_s - 1) % 2 - 1);
-		var t = Math.abs(Math.abs(T_t - 1) % 2 - 1);
-		var x = Math.floor(s * (img.width - 1));
-		var y = Math.floor(t * (img.height - 1));
-		textureSampleLocations.push({x: x, y: y});
-		var d = imageDataGrays[x + y * img.width];
-		inkTextureSamples[i] = d;
-		//samples[i] = 100 + Math.random()*155;
-		
-		// Step texture offset randomly [-1, 1]
-		textureOffsetX += (Math.random() * 2 | 0) === 1 ? -1 : 1;
-		textureOffsetY += (Math.random() * 2 | 0) === 1 ? -1 : 1;
-	}
-
-	return textureSampleLocations;
-}
-
 
 // ------------------------------------------
 // calculateFilteredPoint
@@ -765,12 +710,8 @@ function drawStep(id, point) {
 			if (a < 0) a = 0;
 			if (a >= 1) a = 1;
 
-			// Get new texture sample offset at center
-			var t = inkTextureSamples[textureSampleStep];
-			textureSampleStep = (textureSampleStep === textureSampleLocations.length - 1) ? 0 : (textureSampleStep + 1);
-
 			// Apply texture
-			a *= t;
+			a *= texture.nextInkTextureSample();
 
 			// Grain
 			var g = map(p_p, 0, 1, 0.8, 0.95);
